requirements.txt
# LangChain core
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langchain-chroma>=0.1.0

# Vector store
chromadb==0.5.23

# Scraping
beautifulsoup4==4.12.3
requests>=2.31.0

# Agentes
pyautogen==0.9.0

# UI
streamlit>=1.35.0

# Config y logging
python-dotenv>=1.0.0
pyyaml>=6.0.2
loguru==0.7.3